interface for config

+++ src/conversion/position.hpp
@@ -1,4 +1,3 @@
-
 #include <cstdint>
 
 #pragma once
@@ -7,15 +6,15 @@ class position_converter
 {
 public:
         position_converter(
-            uint16_t lower_value  = 0,
+            uint32_t lower_value  = 0,
             float    lower_angle  = 0.f,
-            uint16_t higher_value = 4096,
+            uint32_t higher_value = 4096,
             float    higher_angle = 3.141592f )
         {
-                int   val_diff   = higher_value - lower_value;
+                int   val_diff   = static_cast< int >( higher_value - lower_value );
                 float angle_diff = higher_angle - lower_angle;
                 scale_           = angle_diff / static_cast< float >( val_diff );
-                offset_          = lower_angle - lower_value * scale_;
+                offset_          = lower_angle - static_cast< float >( lower_value ) * scale_;
         }
 
         float convert( uint32_t val ) const
